% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPmerge_par.R
\name{calc_RMSE}
\alias{calc_RMSE}
\title{Internal function to calculate the Root Mean Square Error to select the maximum interval size.}
\usage{
calc_RMSE(lp_res, max.interval)
}
\arguments{
\item{lp_res}{Result of the mapfuser foreach call or alternatively the result of a call to LPmerge}

\item{max.interval}{A whole number specifying the maximum interval size between bins to included in the objective function for LPmerge}
}
\value{
A list of calculated Root Mean Square Error (RMSE) per chromosome with RMSE for each linkage group ID compared to the consensus map. The mean and standard deviation over the linkage group IDs is also provided.
}
\description{
The LPmerge algorithm is called for all linkage groups with an option for multicore processing. mapfuser further adds automatic selection of the lowest max
}
\author{
Dennis van Muijen
}
