% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_anchors.R
\name{check_anchors}
\alias{check_anchors}
\title{Replace sub-linkage groups (e.g. 1.1, 1.2) with truncated linkage group number and split per chromosome for easy integration}
\usage{
check_anchors(MF.obj = MF.obj, anchors = anchors)
}
\arguments{
\item{MF.obj}{mapfuser object}

\item{anchors}{Number of uniquely positioned anchor markers between genetic maps}
}
\value{
mapfuser object with QC$maps list item containing input maps splitted to unique linkage groups
}
\description{
Replace sub-linkage groups (e.g. 1.1, 1.2) with truncated linkage group number and split per chromosome for easy integration
}
