% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cM.R
\name{check_cM}
\alias{check_cM}
\title{internal position check}
\usage{
check_cM(x)
}
\arguments{
\item{x}{vector of positions}
}
\value{
index of non-numeric items
}
\description{
internal position check
}
\author{
Dennis van Muijen
}
