% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cohesion.R
\name{map_cohesion}
\alias{map_cohesion}
\title{Graph check whether maps can be integrated}
\usage{
map_cohesion(MF.obj)
}
\arguments{
\item{MF.obj}{A mapfuser object}
}
\value{
The input mapfuser object if check OK, error otherwise
}
\description{
Graph check whether maps can be integrated
}
\author{
Dennis van Muijen
}
