% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_flip.R
\name{map_flip}
\alias{map_flip}
\title{Invert a linkage group}
\usage{
map_flip(x)
}
\arguments{
\item{x}{Linkage group to invert}
}
\value{
Linkage group with inverted genetic positions
}
\description{
Invert a linkage group
}
\author{
Dennis van Muijen
}
