% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_orient.R
\name{map_orient}
\alias{map_orient}
\title{Corrects orientation of input genetic maps}
\usage{
map_orient(MF.obj)
}
\arguments{
\item{MF.obj}{A mapfuser object}
}
\value{
A mapfuser object with corrected map orientation
}
\description{
Corrects orientation of input genetic maps
}
\author{
Dennis van Muijen
}
