% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_split.R
\name{map_split}
\alias{map_split}
\title{Replace sub-linkage groups (e.g. 1.1, 1.2) with truncated linkage group number and split per chromosome for easy integration}
\usage{
map_split(MF.obj)
}
\arguments{
\item{MF.obj}{A mapfuser object}
}
\value{
mapfuser object with QC$maps list item containing input maps splitted to unique linkage groups
}
\description{
Replace sub-linkage groups (e.g. 1.1, 1.2) with truncated linkage group number and split per chromosome for easy integration
}
