% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_joinmap.R
\name{read_joinmap}
\alias{read_joinmap}
\title{Internal function, read genetic maps in JoinMap format}
\usage{
read_joinmap(mapfile)
}
\arguments{
\item{mapfile}{JoinMap formatted file to read}
}
\value{
A dataframe with markers, linkage groups identifiers and genetic position on separate columns
}
\description{
Internal function, read genetic maps in JoinMap format
}
\author{
Dennis van Muijen
}
