% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPmerge_par.R
\name{select_result}
\alias{select_result}
\title{Select result from mapfuser}
\usage{
select_result(lp_res, max.int_sel = "auto", map_RMSE, chr = chr)
}
\arguments{
\item{lp_res}{Result of the mapfuser foreach call or alternatively the result of a call to LPmerge}

\item{max.int_sel}{Either automatic selection of the maximum interval size that minimized the mean Root Square Mean Error between the consensus map and invidual linkage maps}

\item{map_RMSE}{Output of a call to calc_RMSE}

\item{chr}{chromosomes to perform analysis for}
}
\value{
The consensus map at the selected maximum interval size K in a convenient format
}
\description{
Internal function for mapfuser -  Flatten, simplify, and combine to one dataframe the consensus map of all linkage groups for easy inspection and exporting
}
\author{
Dennis van Muijen
}
