% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mapfuser.R
\name{summary.mapfuser}
\alias{summary.mapfuser}
\title{internal position check}
\usage{
\method{summary}{mapfuser}(object, ...)
}
\arguments{
\item{object}{mapfuser object}

\item{...}{ingored in function call}
}
\value{
length and number of mapped marker for each input genetic map
}
\description{
internal position check
}
\author{
Dennis van Muijen
}
