% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_export.R
\name{map_export}
\alias{map_export}
\title{Convert a consensus genetic map created with mapfuser to JoinMap format}
\usage{
map_export(MF.obj, file = NULL)
}
\arguments{
\item{MF.obj}{The mapfuser object with filled results slot}

\item{file}{A connection or a string giving the file name to write the map
to.}
}
\description{
Writes the consensus map to the JoinMap ".map" format
}
\examples{
\dontshow{
tmp <- tempfile()
MF.obj <- list()
class(MF.obj) <- "mapfuser"
MF.obj$result$consensus_map <- data.frame(Marker = LETTERS[1:5], LG = c(1,1,1,2,2), Position = 1:5)
map_export(MF.obj, file = tmp)
file.remove(paste0(tmp,".map"))
}
\dontrun{
map_export(MF.obj = MF.obj, file = "consensus")
}
}
\author{
Dennis van Muijen
}
